Move writeFile function to separate utils script file
